requirements.txt
flask
nltk
spacy
requests
beautifulsoup4
numpy
Pillow
wordcloud
tqdm
